requirements.txt
numpy
soundfile
pyloudnorm
pydub
PyYAML
tqdm
pytest
# ffmpeg-python  # optional